auth_server/requirements.txt
Django==2.2.*
djangorestframework
